refactor(test): update the test/ architecture to match the source one
