+++ test/compilation/data/my-external-package.model.json
